+++ tests/php72_files/src/0008_each_deprecated.php
@@ -1,0 +1,6 @@
+<?php
+function iter8(array $arr) {
+    while (list($k, $v) = each($arr)) {
+        echo "$k: $v\n";
+    }
+}