+++ tests/php74_files/src/027_typed_property_mismatch.php
@@ -1,0 +1,10 @@
+<?php
+class A {
+    public $x;
+}
+class B extends A {
+    public int $x;
+}
+class C extends B {
+    public string $x;
+}