+++ tests/plugin_test/src/167_trait_crash.php
@@ -3,7 +3,7 @@
 // Regression test for https://github.com/phan/phan/issues/3796
 namespace TraitCrash;
 /**
- * @method diff(\DateTimeInterface $date)
+ * @method diff(\DateTimeInterface $date) XXX redefining the method as a magic method causes different issue name
  */
 trait BaseTrait {
     public function diff($date = null) {